+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v0.7.x
+- update: error handling for get devices list ([#451](https://github.com/RWTH-EBC/FiLiP/pull/451))
 - add: warning if ``Orion`` version is not supported ([#200](https://github.com/RWTH-EBC/FiLiP/issues/200))
 
 ### v0.7.2